+++ client/src/components/forms/index.js
@@ -5,3 +5,4 @@ export { default as CopyRuleForm } from "./CopyRuleForm";
 export { default as EdgeNetworkingForm } from "./EdgeNetworkingForm";
 export { default as F5Form } from "./F5Form";
 export { default as CbrForm } from "./CbrForm";
+export { default as ObservabilityForm } from "./ObservabilityForm";